Simplify and modernize build of libsncabling
Use simple configure_file and add/target commands to declare sources
and how to build library. No longer copy headers to build directory
as this is a copy-paste from Bayeux, which required that structure due
to its legacy layout of headers.

Alias SNCabling target to "SNCabling::SNCabling" so that superprojects
can include and refer to it as transparently as possible.

Update include paths as needed to find new locations.

Find dependencies directly and use via explicit imported targets.